+++ src/data/data.ts
@@ -1,0 +1,14 @@
+export const features = [
+  {
+    title: "Game mechanic example1",
+    imageSrc: "/images/GM-example1.jpg",
+  },
+  {
+    title: "Game mechanic example2",
+    imageSrc: "/images/GM-example2.jpg",
+  },
+  {
+    title: "Game mechanic example3",
+    imageSrc: "/images/GM-example3.jpg",
+  },
+];
