+++ src/types/globals.ts
@@ -1,14 +1,4 @@
-export const features = [
-  {
-    title: "Game mechanic example1",
-    imageSrc: "/images/GM-example1.jpg",
-  },
-  {
-    title: "Game mechanic example2",
-    imageSrc: "/images/GM-example2.jpg",
-  },
-  {
-    title: "Game mechanic example3",
-    imageSrc: "/images/GM-example3.jpg",
-  },
-];
+export type GameCardProps = {
+  title: string;
+  imageSrc: string;
+};
